docs: record 2026-07-19 audit findings + fixes; lint cleanups
- docs/AUDIT-2026-07-19-fixes.md: full audit outcome (64 confirmed
  findings), the four high-severity fixes with evidence, F24 rebind
  rationale, and the deliberately-deferred items (prelim campaign
  regeneration, Fig. 6 PDF, physicist calls).
- docs/CURRENT-STATUS.md: banner pointing to the follow-up fix branch;
  the snapshot tables still describe pre-fix 71c5f02.
- ruff: ternary in the tau_R corner guard, redundant int(), import sort.

+++ scripts/run_prelim_spatial_finite_phonon_one.py
@@ -141,7 +141,7 @@ def snap_omega_to_grid(omega_uev: float, dE_uev: float) -> tuple[float, int, flo
     """
     if omega_uev <= 0.0 or dE_uev <= 0.0:
         raise ValueError("omega_uev and dE_uev must both be positive.")
-    m = int(round(omega_uev / dE_uev))
+    m = round(omega_uev / dE_uev)
     if m < 1:
         raise ValueError(
             f"omega={omega_uev:g} ueV is below one grid spacing "
